requirements.txt
appnope==0.1.4
asttokens==2.4.1
clarabel==0.9.0
comm==0.2.2
contourpy==1.3.0
cvxpy==1.5.3
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
dgraph_flex==0.1.11
ecos==2.0.14
executing==2.1.0
fastcda==0.1.18
fonttools==4.54.1
graphviz==0.20.3
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.8
jedi==0.19.1
joblib==1.4.2
JPype1==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.16
kiwisolver==1.4.7
lightgbm==4.5.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
miceforest==6.0.3
mpmath==1.3.0
nest-asyncio==1.6.0
numdifftools==0.9.41
numpy==2.1.1
osqp==0.6.7.post3
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.7.0
pyarrow==18.0.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qdldl==0.1.7.post4
scikit-learn==1.5.2
scipy==1.14.1
scs==3.2.7
seaborn==0.13.2
semopy==2.3.11
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.3
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
tzdata==2024.2
wcwidth==0.2.13
widgetsnbextension==4.0.15
